Rename ValidatorExit and move to sdk
