+++ validator/src/dashboard.rs
@@ -6,7 +6,8 @@ use {
     },
     solana_core::validator::ValidatorStartProgress,
     solana_sdk::{
-        clock::Slot, commitment_config::CommitmentConfig, native_token::Sol, pubkey::Pubkey,
+        clock::Slot, commitment_config::CommitmentConfig, exit::Exit, native_token::Sol,
+        pubkey::Pubkey,
     },
     std::{
         io,
@@ -31,7 +32,7 @@ impl Dashboard {
     pub fn new(
         ledger_path: &Path,
         log_path: Option<&Path>,
-        validator_exit: Option<&mut solana_core::validator::ValidatorExit>,
+        validator_exit: Option<&mut Exit>,
     ) -> Result<Self, io::Error> {
         println_name_value("Ledger location:", &format!("{}", ledger_path.display()));
         if let Some(log_path) = log_path {
